+++ src/Core/src/Attribute/Scope.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Spiral\Core\Attribute;
 
 /**
- * Set a scope in which the dependency can be resolved.
+ * Set the scope in which the dependency can be resolved.
  *
  * @internal We are testing this feature, it may be changed in the future.
  */