+++ src/Core/src/Internal/Factory.php
@@ -8,12 +8,9 @@ use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
 use ReflectionFunctionAbstract as ContextFunction;
 use ReflectionParameter;
-use Spiral\Core\Attribute\Finalize;
-use Spiral\Core\Attribute\Scope as ScopeAttribute;
-use Spiral\Core\Attribute\Singleton;
+use Spiral\Core\Attribute;
 use Spiral\Core\BinderInterface;
-use Spiral\Core\Config\Injectable;
-use Spiral\Core\Config\DeferredFactory;
+use Spiral\Core\Config;
 use Spiral\Core\Container\InjectorInterface;
 use Spiral\Core\Container\SingletonInterface;
 use Spiral\Core\Exception\Container\AutowireException;
@@ -29,6 +26,7 @@ use Spiral\Core\Internal\Common\DestructorTrait;
 use Spiral\Core\Internal\Common\Registry;
 use Spiral\Core\Internal\Factory\Ctx;
 use Spiral\Core\InvokerInterface;
+use Spiral\Core\Options;
 use Spiral\Core\ResolverInterface;
 use Stringable;
 use WeakReference;
@@ -47,6 +45,7 @@ final class Factory implements FactoryInterface
     private ResolverInterface $resolver;
     private Tracer $tracer;
     private Scope $scope;
+    private Options $options;
 
     public function __construct(Registry $constructor)
     {
@@ -59,6 +58,7 @@ final class Factory implements FactoryInterface
         $this->resolver = $constructor->get('resolver', ResolverInterface::class);
         $this->tracer = $constructor->get('tracer', Tracer::class);
         $this->scope = $constructor->get('scope', Scope::class);
+        $this->options = $constructor->getOptions();
     }
 
     /**
@@ -91,18 +91,20 @@ final class Factory implements FactoryInterface
 
             unset($this->state->bindings[$alias]);
             return match ($binding::class) {
-                \Spiral\Core\Config\Alias::class => $this->resolveAlias($binding, $alias, $context, $parameters),
-                \Spiral\Core\Config\Autowire::class => $this->resolveAutowire($binding, $alias, $context, $parameters),
-                DeferredFactory::class,
-                \Spiral\Core\Config\Factory::class => $this->resolveFactory($binding, $alias, $context, $parameters),
-                \Spiral\Core\Config\Shared::class => $this->resolveShared($binding, $alias, $context, $parameters),
-                Injectable::class => $this->resolveInjector(
+                Config\Alias::class => $this->resolveAlias($binding, $alias, $context, $parameters),
+                Config\Proxy::class,
+                Config\DeprecationProxy::class => $this->resolveProxy($binding, $alias, $context),
+                Config\Autowire::class => $this->resolveAutowire($binding, $alias, $context, $parameters),
+                Config\DeferredFactory::class,
+                Config\Factory::class => $this->resolveFactory($binding, $alias, $context, $parameters),
+                Config\Shared::class => $this->resolveShared($binding, $alias, $context, $parameters),
+                Config\Injectable::class => $this->resolveInjector(
                     $binding,
                     new Ctx(alias: $alias, class: $alias, context: $context),
                     $parameters,
                 ),
-                \Spiral\Core\Config\Scalar::class => $binding->value,
-                \Spiral\Core\Config\WeakReference::class => $this
+                Config\Scalar::class => $binding->value,
+                Config\WeakReference::class => $this
                     ->resolveWeakReference($binding, $alias, $context, $parameters),
                 default => $binding,
             };
@@ -117,7 +119,7 @@ final class Factory implements FactoryInterface
      * @psalm-suppress UnusedParam
      * todo wat should we do with $arguments?
      */
-    private function resolveInjector(Injectable $binding, Ctx $ctx, array $arguments)
+    private function resolveInjector(Config\Injectable $binding, Ctx $ctx, array $arguments)
     {
         $context = $ctx->context;
         try {
@@ -174,7 +176,7 @@ final class Factory implements FactoryInterface
     }
 
     private function resolveAlias(
-        \Spiral\Core\Config\Alias $binding,
+        Config\Alias $binding,
         string $alias,
         Stringable|string|null $context,
         array $arguments,
@@ -194,8 +196,19 @@ final class Factory implements FactoryInterface
         return $result;
     }
 
+    private function resolveProxy(Config\Proxy $binding, string $alias, Stringable|string|null $context): mixed
+    {
+        $result = Proxy::create(new \ReflectionClass($binding->getInterface()), $context, new Attribute\Proxy());
+
+        if ($binding->singleton) {
+            $this->state->singletons[$alias] = $result;
+        }
+
+        return $result;
+    }
+
     private function resolveShared(
-        \Spiral\Core\Config\Shared $binding,
+        Config\Shared $binding,
         string $alias,
         Stringable|string|null $context,
         array $arguments,
@@ -210,7 +223,7 @@ final class Factory implements FactoryInterface
     }
 
     private function resolveAutowire(
-        \Spiral\Core\Config\Autowire $binding,
+        Config\Autowire $binding,
         string $alias,
         Stringable|string|null $context,
         array $arguments,
@@ -222,14 +235,14 @@ final class Factory implements FactoryInterface
     }
 
     private function resolveFactory(
-        \Spiral\Core\Config\Factory|DeferredFactory $binding,
+        Config\Factory|Config\DeferredFactory $binding,
         string $alias,
         Stringable|string|null $context,
         array $arguments,
     ): mixed {
         $ctx = new Ctx(alias: $alias, class: $alias, context: $context, singleton: $binding->singleton);
         try {
-            $instance = $binding::class === \Spiral\Core\Config\Factory::class && $binding->getParametersCount() === 0
+            $instance = $binding::class === Config\Factory::class && $binding->getParametersCount() === 0
                 ? ($binding->factory)()
                 : $this->invoker->invoke($binding->factory, $arguments);
         } catch (NotCallableException $e) {
@@ -244,7 +257,7 @@ final class Factory implements FactoryInterface
     }
 
     private function resolveWeakReference(
-        \Spiral\Core\Config\WeakReference $binding,
+        Config\WeakReference $binding,
         string $alias,
         Stringable|string|null $context,
         array $arguments,
@@ -286,7 +299,7 @@ final class Factory implements FactoryInterface
         array $parameters = [],
         Stringable|string|null $context = null
     ): mixed {
-        $parent = $this->scope->getParent();
+        $parent = $this->scope->getParentFactory();
 
         if ($parent !== null) {
             try {
@@ -294,6 +307,7 @@ final class Factory implements FactoryInterface
                     'current scope' => $this->scope->getScopeName(),
                     'jump to parent scope' => $this->scope->getParentScope()->getScopeName(),
                 ]);
+                /** @psalm-suppress TooManyArguments */
                 return $parent->make($alias, $parameters, $context);
             } catch (BadScopeException $e) {
                 if ($this->scope->getScopeName() !== $e->getScope()) {
@@ -367,11 +381,16 @@ final class Factory implements FactoryInterface
         Ctx $ctx,
         array $arguments,
     ): object {
-        // Check scope name
-        $ctx->reflection = new \ReflectionClass($instance);
-        $scopeName = ($ctx->reflection->getAttributes(ScopeAttribute::class)[0] ?? null)?->newInstance()->name;
-        if ($scopeName !== null && $scopeName !== $this->scope->getScopeName()) {
-            throw new BadScopeException($scopeName, $instance::class);
+        if ($this->options->checkScope) {
+            // Check scope name
+            $ctx->reflection = new \ReflectionClass($instance);
+            $scopeName = ($ctx->reflection->getAttributes(Attribute\Scope::class)[0] ?? null)?->newInstance()->name;
+            if ($scopeName !== null) {
+                $scope = $this->scope;
+                while ($scope->getScopeName() !== $scopeName) {
+                    $scope = $scope->getParentScope() ?? throw new BadScopeException($scopeName, $instance::class);
+                }
+            }
         }
 
         return $arguments === []
@@ -404,9 +423,11 @@ final class Factory implements FactoryInterface
         }
 
         // Check scope name
-        $scope = ($reflection->getAttributes(ScopeAttribute::class)[0] ?? null)?->newInstance()->name;
-        if ($scope !== null && $scope !== $this->scope->getScopeName()) {
-            throw new BadScopeException($scope, $class);
+        if ($this->options->checkScope) {
+            $scope = ($reflection->getAttributes(Attribute\Scope::class)[0] ?? null)?->newInstance()->name;
+            if ($scope !== null && $scope !== $this->scope->getScopeName()) {
+                throw new BadScopeException($scope, $class);
+            }
         }
 
         // We have to construct class using external injector when we know the exact context
@@ -431,7 +452,7 @@ final class Factory implements FactoryInterface
             try {
                 $this->tracer->push(false, action: 'resolve arguments', signature: $constructor);
                 $this->tracer->push(true);
-                $args = $this->resolver->resolveArguments($constructor, $arguments);
+                $args = $this->resolver->resolveArguments($constructor, $arguments, $this->options->validateArguments);
             } catch (ValidationException $e) {
                 throw new ContainerException(
                     $this->tracer->combineTraceMessage(
@@ -502,16 +523,16 @@ final class Factory implements FactoryInterface
             return true;
         }
 
-        return $ctx->reflection->getAttributes(Singleton::class) !== [];
+        return $ctx->reflection->getAttributes(Attribute\Singleton::class) !== [];
     }
 
     private function getFinalizer(Ctx $ctx, object $instance): ?callable
     {
         /**
          * @psalm-suppress UnnecessaryVarAnnotation
-         * @var Finalize|null $attribute
+         * @var Attribute\Finalize|null $attribute
          */
-        $attribute = ($ctx->reflection->getAttributes(Finalize::class)[0] ?? null)?->newInstance();
+        $attribute = ($ctx->reflection->getAttributes(Attribute\Finalize::class)[0] ?? null)?->newInstance();
         if ($attribute === null) {
             return null;
         }