+++ src/Core/tests/Internal/Proxy/ProxyTest.php
@@ -4,9 +4,14 @@ declare(strict_types=1);
 
 namespace Spiral\Tests\Core\Internal\Proxy;
 
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\WithoutErrorHandler;
 use PHPUnit\Framework\TestCase;
 use Spiral\Core\Attribute\Proxy;
+use Spiral\Core\Config;
 use Spiral\Core\Container;
+use Spiral\Core\Exception\Container\ContainerException;
+use Spiral\Core\Scope;
 use Spiral\Tests\Core\Internal\Proxy\Stub\EmptyInterface;
 use Spiral\Tests\Core\Internal\Proxy\Stub\MockInterface;
 use Spiral\Tests\Core\Internal\Proxy\Stub\MockInterfaceImpl;
@@ -20,9 +25,7 @@ final class ProxyTest extends TestCase
         // yield [EmptyInterface::class, 'empty'];
     }
 
-    /**
-     * @dataProvider interfacesProvider
-     */
+    #[DataProvider('interfacesProvider')]
     public function testSimpleCases(string $interface, string $var): void
     {
         $root = new Container();
@@ -54,9 +57,7 @@ final class ProxyTest extends TestCase
         });
     }
 
-    /**
-     * @dataProvider interfacesProvider
-     */
+    #[DataProvider('interfacesProvider')]
     public function testExtraArguments(string $interface, string $var): void
     {
         $root = new Container();
@@ -76,9 +77,7 @@ final class ProxyTest extends TestCase
         });
     }
 
-    /**
-     * @dataProvider interfacesProvider
-     */
+    #[DataProvider('interfacesProvider')]
     public function testVariadic(string $interface, string $var): void
     {
         $root = new Container();
@@ -99,9 +98,7 @@ final class ProxyTest extends TestCase
         });
     }
 
-    /**
-     * @dataProvider interfacesProvider
-     */
+    #[DataProvider('interfacesProvider')]
     public function testReference(string $interface, string $var): void
     {
         $interface === EmptyInterface::class && self::markTestSkipped(
@@ -121,9 +118,7 @@ final class ProxyTest extends TestCase
         });
     }
 
-    /**
-     * @dataProvider interfacesProvider
-     */
+    #[DataProvider('interfacesProvider')]
     public function testReturnReference(string $interface, string $var): void
     {
         $interface === EmptyInterface::class && self::markTestSkipped(
@@ -146,9 +141,7 @@ final class ProxyTest extends TestCase
         });
     }
 
-    /**
-     * @dataProvider interfacesProvider
-     */
+    #[DataProvider('interfacesProvider')]
     public function testReferenceVariadic(string $interface, string $var): void
     {
         $interface === EmptyInterface::class && self::markTestSkipped(
@@ -187,5 +180,159 @@ final class ProxyTest extends TestCase
         $mock = $root->invoke(static fn(#[Proxy] MockInterface|EmptyInterface $mock) => $mock);
 
         self::assertInstanceOf(MockInterfaceImpl::class, $mock);
+    }
+
+    #[DataProvider('interfacesProvider')]
+    public function testProxyConfig(string $interface): void
+    {
+        $root = new Container();
+        $root->getBinder('foo')->bindSingleton($interface, Stub\MockInterfaceImpl::class);
+        $root->bindSingleton($interface, new Config\Proxy($interface, true));
+
+        $proxy = $root->get($interface);
+        $this->assertInstanceOf($interface, $proxy);
+        $this->assertNotInstanceOf(MockInterfaceImpl::class, $proxy);
+
+        $root->runScope(new Scope('foo'), static function (Container $container) use ($interface, $proxy) {
+            $proxy->bar(name: 'foo'); // Possible to run
+            self::assertSame('foo', $proxy->baz('foo', 42));
+            self::assertSame(123, $proxy->qux(age: 123));
+            self::assertSame(69, $proxy->space(test age: 69));
+
+            $real = $container->get($interface);
+            self::assertInstanceOf(MockInterfaceImpl::class, $real);
+
+            $real->bar(name: 'foo'); // Possible to run
+            self::assertSame('foo', $real->baz('foo', 42));
+            self::assertSame(123, $real->qux(age: 123));
+            self::assertSame(69, $real->space(test age: 69));
+        });
+    }
+
+    #[DataProvider('interfacesProvider')]
+    public function testProxyConfigOutOfProxyException(string $interface): void
+    {
+        $root = new Container();
+        $root->getBinder('foo')->bindSingleton($interface, Stub\MockInterfaceImpl::class);
+        $root->bindSingleton($interface, new Config\Proxy($interface, true));
+
+        $this->assertInstanceOf($interface, $root->get($interface));
+        $proxy = $root->get($interface);
+
+        $this->expectException(ContainerException::class);
+        $this->expectExceptionMessage('Proxy is out of scope.');
+        $proxy->bar(name: 'foo'); // Impossible to run
+    }
+
+    #[DataProvider('interfacesProvider')]
+    #[WithoutErrorHandler]
+    public function testDeprecationProxyConfig(string $interface): void
+    {
+        \set_error_handler(static function (int $errno, string $error) use ($interface): void {
+            self::assertSame(
+                \sprintf('Using `%s` outside of the `foo` scope is deprecated and will be ' .
+                    'impossible in version 4.0.', $interface),
+                $error
+            );
+        });
+
+        $root = new Container();
+        $root->getBinder('foo')->bindSingleton($interface, Stub\MockInterfaceImpl::class);
+        $root->bindSingleton($interface, new Config\DeprecationProxy($interface, true, 'foo', '4.0'));
+
+        $proxy = $root->get($interface);
+        $this->assertInstanceOf($interface, $proxy);
+
+        $root->runScope(new Scope('foo'), static function () use ($proxy) {
+            $proxy->bar(name: 'foo'); // Possible to run
+            self::assertSame('foo', $proxy->baz('foo', 42));
+            self::assertSame(123, $proxy->qux(age: 123));
+            self::assertSame(69, $proxy->space(test age: 69));
+        });
+
+        \restore_error_handler();
+    }
+
+    #[DataProvider('interfacesProvider')]
+    #[WithoutErrorHandler]
+    public function testDeprecationProxyConfigCustomMessage(string $interface): void
+    {
+        \set_error_handler(static function (int $errno, string $error) use ($interface): void {
+            self::assertSame(\sprintf('Using `%s` impossible', $interface), $error);
+        });
+
+        $root = new Container();
+        $root->getBinder('foo')->bindSingleton($interface, Stub\MockInterfaceImpl::class);
+        $root->bindSingleton($interface, new Config\DeprecationProxy(
+            interface: $interface,
+            message: \sprintf('Using `%s` impossible', $interface),
+        ));
+
+        $proxy = $root->get($interface);
+        $this->assertInstanceOf($interface, $proxy);
+
+        $root->runScope(new Scope('foo'), static function () use ($proxy) {
+            $proxy->bar(name: 'foo'); // Possible to run
+            self::assertSame('foo', $proxy->baz('foo', 42));
+            self::assertSame(123, $proxy->qux(age: 123));
+            self::assertSame(69, $proxy->space(test age: 69));
+        });
+
+        \restore_error_handler();
+    }
+
+    #[DataProvider('interfacesProvider')]
+    #[WithoutErrorHandler]
+    public function testDeprecationProxyConfigDontThrowIfNotConstructed(string $interface): void
+    {
+        \set_error_handler(static function (int $errno, string $error) use ($interface): void {
+            self::fail('Unexpected error: ' . $error);
+        });
+
+        $root = new Container();
+        $root->getBinder('foo')->bindSingleton($interface, Stub\MockInterfaceImpl::class);
+        $root->bindSingleton($interface, new Config\DeprecationProxy($interface, true, 'foo', '4.0'));
+
+        $root->runScope(new Scope('foo'), static function (Container $container) use ($interface) {
+            $proxy = $container->get($interface);
+
+            $proxy->bar(name: 'foo'); // Possible to run
+            self::assertSame('foo', $proxy->baz('foo', 42));
+            self::assertSame(123, $proxy->qux(age: 123));
+            self::assertSame(69, $proxy->space(test age: 69));
+        });
+
+        \restore_error_handler();
+    }
+
+    #[DataProvider('invalidDeprecationProxyArgsDataProvider')]
+    public function testDeprecationProxyConfigArgsRequiredException(string|null $scope, string|null $version): void
+    {
+        self::expectException(\InvalidArgumentException::class);
+        self::expectExceptionMessage('Scope and version or custom message must be provided.');
+
+        new Config\DeprecationProxy(interface: EmptyInterface::class, scope: $scope, version: $version);
+    }
+
+    public function testProxyConfigToString(): void
+    {
+        $proxy = new Config\Proxy(EmptyInterface::class);
+
+        $this->assertSame(\sprintf('Proxy to `%s`', EmptyInterface::class), (string) $proxy);
+    }
+
+    public function testProxyConfigNotInterfaceException(): void
+    {
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage(\sprintf('Interface `%s` does not exist.', \stdClass::class));
+
+        new Config\Proxy(\stdClass::class);
+    }
+
+    public static function invalidDeprecationProxyArgsDataProvider(): \Traversable
+    {
+        yield [null, '4.0'];
+        yield ['foo', null];
+        yield [null, null];
     }
 }